pyproject.toml
[tool.poetry]
name = "wellbex-test"
version = "0.1.0"
description = "wellbex-test"
authors = ["Artem Merkulov <artem-merk96@yandex.ru>"]
readme = "README.md"
packages = [{include = "wellbex_test"}]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.0"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.1"
psycopg2-binary = "^2.9.9"
envparse = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
